Consolidate mailer actions into tasks
