+++ communication/messages.py
@@ -1,36 +1,30 @@
 from http import HTTPStatus
-from typing import List, Set, Tuple
+from typing import List, Tuple
 
 from fastapi import APIRouter, Depends, HTTPException
-from mailer import AsyncClient, BodyType
 from opentelemetry import trace
 from pydantic import validate_model
 from sqlalchemy import select
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
-from sqlalchemy.sql import Select, or_
 
-from common import SETTINGS
 from common.authentication import with_user_id
 from common.database import (
-    Application,
     Group,
     Message,
     MessageCreate,
     MessageList,
     MessageRead,
     MessageUpdate,
-    Participant,
     Recipient,
     RecipientCreate,
     RecipientRead,
     with_db,
 )
-from common.mail import MJMLClient, with_mail, with_mjml
+from common.mjml import MJMLClient, with_mjml
 from common.permissions import Permission, requires_permission
-
-from .util import send_message
+from common.tasks import tasks
 
 router = APIRouter()
 tracer = trace.get_tracer(__name__)
@@ -212,7 +206,6 @@ async def delete_recipient(id: int, group: Group, db: AsyncSession = Depends(wit
 )
 async def send(
     id: int,
-    mailer: AsyncClient = Depends(with_mail),
     db: AsyncSession = Depends(with_db),
 ):
     """
@@ -222,27 +215,7 @@ async def send(
     if message is None:
         raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="not found")
 
-    # Retrieve all the recipient emails
-    groups = {r.group for r in message.recipients}
-    result = await db.execute(recipients_query(groups))
-    participants = result.all()
-
-    with tracer.start_as_current_span("build-context"):
-        emails = {}
-        for row in participants:
-            emails[row[2]] = {"first_name": row[0], "last_name": row[1]}
-
-    await mailer.send_template(
-        emails,
-        SETTINGS.communication.sender,
-        message.subject,
-        message.rendered,
-        body_type=BodyType.HTML if message.is_html else BodyType.PLAIN,
-        reply_to=SETTINGS.communication.reply_to,
-    )
-
-    message.sent = True
-    await db.commit()
+    await tasks.communication.send(message_id=id)
 
 
 @router.post(
@@ -254,7 +227,6 @@ async def send(
 async def send_test(
     id: int,
     user_id: str = Depends(with_user_id),
-    mailer: AsyncClient = Depends(with_mail),
     db: AsyncSession = Depends(with_db),
 ):
     """
@@ -264,11 +236,7 @@ async def send_test(
     if message is None:
         raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="not found")
 
-    # Get the current user
-    user = await db.get(Participant, user_id)
-    assert user is not None
-
-    await send_message(user, message, mailer)
+    await tasks.communication.send_test(message_id=message.id, user_id=user_id)
 
 
 @router.delete(
@@ -285,40 +253,6 @@ async def delete(id: int, db: AsyncSession = Depends(with_db)):
     if message:
         await db.delete(message)
         await db.commit()
-
-
-def recipients_query(groups: Set[Group]) -> Select:
-    """
-    Generate the query for recipients when sending emails. Also performs some simple optimizations.
-    :param groups: the recipient groups for the message
-    :returns: a SQL query
-    """
-    fields = [Participant.first_name, Participant.last_name, Participant.email]
-
-    # All application statues and all application completion states can be simplified to everyone
-    if (
-        Group.EVERYONE in groups
-        or groups == Group.completion_states()
-        or groups == Group.statuses()
-    ):
-        return select(*fields)
-
-    status_filter = Application.status.in_(  # type: ignore
-        [g.to_status() for g in groups if g.to_status() is not None]
-    )
-    base = select(*fields).outerjoin(Application, full=True)
-
-    # If filtering for complete, status does not matter
-    if Group.APPLICATION_COMPLETE in groups:
-        return base.where(Application.participant_id != None)
-
-    # Add extra filter for incomplete
-    elif Group.APPLICATION_INCOMPLETE in groups:
-        return base.where(or_(status_filter, Application.participant_id == None))
-
-    # Only filter on status
-    else:
-        return base.where(status_filter)
 
 
 async def render_mjml(source: str, client: MJMLClient) -> Tuple[str, bool]:
